add separate test module
