+++ src/service/mod.rs
@@ -1,1 +1,2 @@
 pub mod auth_service;
+mod tests;
